pyproject.toml
[tool.poetry]
name = "pubmed-pharma-paper-finder"
version = "0.1.0"
description = "A tool to find PubMed papers with authors affiliated to pharmaceutical companies"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
repository = "https://github.com/yourusername/pubmed-pharma-paper-finder"
packages = [{include = "pubmed_pharma_paper_finder"}]

[tool.poetry.dependencies]
python = "^3.12"
biopython = "^1.81"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
mypy = "^1.5.1"
isort = "^5.12.0"
flake8 = "^7.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
get-papers-list = "pubmed_pharma_paper_finder.cli:main"
